move DevSupportManagerFactory out from interfaces/ to fix buck dependency issue
