+++ ReactAndroid/src/main/java/com/facebook/react/ReactNativeHost.java
@@ -16,7 +16,7 @@ import com.facebook.react.bridge.ReactMarker;
 import com.facebook.react.bridge.ReactMarkerConstants;
 import com.facebook.react.common.LifecycleState;
 import com.facebook.react.devsupport.RedBoxHandler;
-import com.facebook.react.devsupport.interfaces.DevSupportManagerFactory;
+import com.facebook.react.devsupport.DevSupportManagerFactory;
 import com.facebook.react.uimanager.UIImplementationProvider;
 import java.util.List;
