+++ ReactAndroid/src/main/java/com/facebook/react/devsupport/DefaultDevSupportManagerFactory.java
@@ -11,7 +11,6 @@ import android.content.Context;
 import androidx.annotation.Nullable;
 import com.facebook.react.devsupport.interfaces.DevBundleDownloadListener;
 import com.facebook.react.devsupport.interfaces.DevSupportManager;
-import com.facebook.react.devsupport.interfaces.DevSupportManagerFactory;
 import com.facebook.react.packagerconnection.RequestHandler;
 import java.lang.reflect.Constructor;
 import java.util.Map;
